lc_lib Update to LeoCAD 1952 mLCStatusBar     KO mStatusBarLabel  KO mStatusSnapLable KO

+++ lc_lib/common/image.cpp
@@ -149,9 +149,9 @@ bool Image::FileLoad(lcMemFile& File)
 	QImage Image;
 
 	unsigned char* Buffer = File.mBuffer + File.mPosition;
-	int BufferLength = File.mFileSize - File.mPosition;
+	size_t BufferLength = File.mFileSize - File.mPosition;
 
-	if (!Image.loadFromData(Buffer, BufferLength))
+	if (!Image.loadFromData(Buffer, (int)BufferLength))
 		return false;
 
 	CopyFromQImage(Image, *this);
